[Monterey] LayoutTest media/element-containing-pip-video-going-into-fullscreen.html is flaky timeout/crash https://bugs.webkit.org/show_bug.cgi?id=229453 <rdar://80346428>
Reviewed by Eric Carlson.

Source/WebCore:

No new tests. Fix a flaky test.

* html/HTMLMediaElement.cpp:
(WebCore::HTMLMediaElement::HTMLMediaElement):
We need to initialize `m_changingVideoFullscreenMode` to false in the constructor.
Otherwise, it might be initialized to true and ignore requests to change video
presentation mode.

Source/WebKit:

* UIProcess/Cocoa/VideoFullscreenManagerProxy.mm:
(WebKit::VideoFullscreenManagerProxy::exitFullscreenWithoutAnimationToMode):
When the mock video presentation mode is enabled, the UI process still needs to
notify the Web process that the mode change has been done.

* WebProcess/cocoa/VideoFullscreenManager.mm:
(WebKit::VideoFullscreenManager::didCleanupFullscreen):
Return immediately if the element is not in fullscreen to avoid
an assertion later in `enterVideoFullscreenForVideoElement()`.

LayoutTests:

* media/element-containing-pip-video-going-into-fullscreen.html:
Enable the mock video presentation mode for testing.
* platform/mac-wk2/TestExpectations:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@281557 268f45cc-cd09-0410-ab3c-d52691b4dbfc