[IFC] Use the inline run list in showRenderTree to print inline level box information https://bugs.webkit.org/show_bug.cgi?id=228070
Reviewed by Antti Koivisto.

showInlineTreeAndRuns is the last client of the InlineFormattingState::lineBoxes.
Let's use the inline run list instead to print the inline level box geometry information. It provides slightly less
information than before but it is sufficient for now.

This is in preparation for keeping the inline box structure private to the layout code.

* layout/layouttree/LayoutTreeBuilder.cpp:
(WebCore::Layout::showInlineTreeAndRuns):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@281645 268f45cc-cd09-0410-ab3c-d52691b4dbfc