[LFC][Integration] Use the line runs to check for legacy integral positioning https://bugs.webkit.org/show_bug.cgi?id=228069 <rdar://problem/81087857>
Reviewed by Antti Koivisto.

Now that the line runs list has all the inline boxes (including spanning inline boxes), we
can use this list to check for legacy integral positioning and not directly iterating the LineBox's non-rootinlinebox list.
While this may be a bit more expensive, the legacy positioning is temporary and will be removed shortly.

This is in preparation for keeping the inline box structure private to the layout code.

* layout/integration/LayoutIntegrationInlineContentBuilder.cpp:
(WebCore::LayoutIntegration::InlineContentBuilder::build const):
(WebCore::LayoutIntegration::InlineContentBuilder::computeLineLevelVisualAdjustmentsForRuns const):
* layout/integration/LayoutIntegrationInlineContentBuilder.h:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@281680 268f45cc-cd09-0410-ab3c-d52691b4dbfc