[Cocoa] Enforce the policy against WebKit causing Core Text font download prompts in WebKit1 https://bugs.webkit.org/show_bug.cgi?id=229391
Reviewed by Alan Bujtas.

Source/WebCore:

For most apps, if you ask Core Text to create a font which is a mobileasset font, it will
show a synchronous blocking prompt asking the user if they want to download the font.
However, this policy isn't a great one for web content, because webpages will request tons
of fonts all the time, and it's totally likely that many of them will be mobileasset fonts.
Because of this, WebKit has a policy where we will opt out of showing these prompts, by
specifying kCTFontEnabledAttribute:kCFBooleanTrue in the attributes dictionary.

We do this in some places, but we forgot one additional place. This patch adds it to this
place, and adds a test which intentionally requests a MobileAsset font and makes sure the
test doesn't timeout (by showing a blocking prompt forever).

Test: fast/text/mobileasset-font.html

* platform/graphics/cocoa/FontCacheCoreText.cpp:
(WebCore::autoActivateFont):

LayoutTests:

Make sure content which uses these fonts doesn't cause a blocking prompt.

* fast/text/mobileasset-font-expected-mismatch.html: Added.
* fast/text/mobileasset-font.html: Added.
* platform/mac/TestExpectations: Skip the test on Catalina, because apparently the prompts can't be skipped?


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@281681 268f45cc-cd09-0410-ab3c-d52691b4dbfc