Caret should respect text background color https://bugs.webkit.org/show_bug.cgi?id=117493
Patch by Joone Hur <joone@webkit.org> on 2021-08-26
Reviewed by Ryosuke Niwa.

Source/WebCore:

An earlier fix for caret color, which we are now improving, was this commit:
https://trac.webkit.org/changeset/152612/webkit

This patch allows the caret to become visible in the black background
by getting the caret color from the element containing
the text, not the root editable element that has the contentEditable attribute.

* editing/FrameSelection.cpp:
(WebCore::CaretBase::computeCaretColor):

LayoutTests:

Rebaselined caret rendering tests.

* editing/caret/caret-color.html:
* editing/caret/color-span-inside-editable-background-expected.html:
* editing/caret/color-span-inside-editable-expected.html:
* fast/css/caret-color-auto-expected.html:
* platform/ios-wk2/editing/caret/caret-color-expected.png:
* platform/ios/editing/caret/caret-color-expected.txt:
* platform/mac/editing/caret/caret-color-expected.png:
* platform/mac/editing/caret/caret-color-expected.txt:
* platform/win/editing/caret/caret-color-expected.txt:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@281685 268f45cc-cd09-0410-ab3c-d52691b4dbfc