FontFaceSet.status needs to react to style changes https://bugs.webkit.org/show_bug.cgi?id=229839
Reviewed by Sam Weinig.

Source/WebCore:

When content says "document.fonts.status", the ".status" part needs to
update style instead of the "document.fonts" part. This is because it's
totally legal for content to say:

var f = document.fonts;
// Modify style here
... f.status ... <=== This needs to reflect the style changes.

Test: fast/text/FontFaceSet-status-after-style-update.html

* css/FontFaceSet.cpp:
(WebCore::FontFaceSet::status const):

LayoutTests:

This is not a WPT test because it requires internals.invalidateFontCache().

* fast/text/FontFaceSet-status-after-style-update-expected.txt: Added.
* fast/text/FontFaceSet-status-after-style-update.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@282015 268f45cc-cd09-0410-ab3c-d52691b4dbfc