@font-face rules accessed over file: urls to a file which does not exist are not visible to document.fonts https://bugs.webkit.org/show_bug.cgi?id=229639
Reviewed by Alan Bujtas.

LayoutTests/imported/w3c:

* web-platform-tests/css/css-font-loading/nonexistent-file-url-expected.txt:

Source/WebCore:

Because all the sources were failed, we were marking the whole CSSFontFace as failed, and then
not adding it to the CSSFontFaceSource. Unfortunately, this optimization is observable, so we
can't do it.

Tests: fast/text/font-file-failed.html
       imported/w3c/web-platform-tests/css/css-font-loading/nonexistent-file-url.html

* css/CSSFontSelector.cpp:
(WebCore::CSSFontSelector::addFontFaceRule):

LayoutTests:

Keep a local copy of these tests, because WPT tests don't run from file: URLs, but layout tests do.
The behavior is different depending on the scheme of the main page.

* fast/text/font-file-failed-expected.txt: Added.
* fast/text/font-file-failed.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@282017 268f45cc-cd09-0410-ab3c-d52691b4dbfc