Ensure fragmented flow state invalidation even when the cached fragmented flow is not present. https://bugs.webkit.org/show_bug.cgi?id=229914 <rdar://82025006>
Reviewed by Antti Koivisto.

The cached fragmented flow is the byproduct of querying the enclosing fragment during layout/repaint.
Sometimes when the layout process, between 2 subsequent style change does not generate such cached fragmented flows
the invalidation process stops early and leaves subtrees in an inconsistent state.

Let's use the passed in fragmented flow when the cached value is not present.

* rendering/RenderBlock.cpp:
(WebCore::RenderBlock::resetEnclosingFragmentedFlowAndChildInfoIncludingDescendants):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@282047 268f45cc-cd09-0410-ab3c-d52691b4dbfc