Remove unnecessary uses of Ref protect = m_backing from FontFaceSet https://bugs.webkit.org/show_bug.cgi?id=229911
Reviewed by Anders Carlsson.

The only callers of these are from JS. So, JS will maintain a ref to
the FontFaceSet, and the FontFaceSet will maintain a ref to its
CSSFontFaceSet. So it doesn't need to be protected.

No new tests because there is no behavior change.

* css/FontFaceSet.cpp:
(WebCore::FontFaceSet::size):
(WebCore::FontFaceSet::check):
(WebCore::FontFaceSet::status const):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@282049 268f45cc-cd09-0410-ab3c-d52691b4dbfc